Includes new import and export to file methods for Pokemon bytes, as well as proper copying file methods to box slots. This revision should be able to write a pokemon to a file using getPosition(2,1) (box 2, slot1) and then write it to another position, in either the same or a different game's box. writeFile should be called whenever writePokemonToData is.
The format of a pokemon's exported filename is (Nickname) the (Species)
((checksum)).2pkm. The checksum is mostly arbitrary as it is only
relevant to this program, not the game.

Future things: proper party importing (needs to adjust the mirror copy
as well), box nickname reading, further testing, gold and silver
support, and then the rest of the generations

tldr Coming along well! Samples of the exported files are included.